+++ includes/OpenLevelController.hpp
@@ -59,9 +59,9 @@ public:
      *
      * @param pContext commons items for controller
      *
-     * @return short   id of the next screen controller
+     * @return uint8_t   id of the next screen controller
      */
-    unsigned short render(utils::Context* pContext);
+    uint8_t render(utils::Context* pContext);
 
 private:
